executor: migrate test-infra to testify for two parts of executor_test
